Refactor template package.json saving
